Add Schema Filtering to Postgres dbs.
Adjust the Postgres Driver to add the Schema filtering settings to the UI in the admin page.

Fix a bug where "all" schema filter case is ignored, defaulting (incorrectly) to use only the "public" schema. This
would lead to potentially filtering away schemas that the user expects, if some tables exist inside schemas that
aren't public.